Redesign, part 3.
Experimental redesign work: Add interpreter and renderer
modules to replace the generator module. (The interpreter
is to do all program running work which is separate from
producing audio; the renderer becoming the new (and simpler)
sound generator.) Work on these is unfinished; a #define
TEST_INTERPRETER switch is added to sgensys.h to use them
instead of the old sound generator. Disabled by default.

Based on old redesign work from 2013 and 2014, discarding
the old changes for program module and parser.c. (The
redesign of the sound generator is based on making it
similar to before graph structure for voice & op. was
introduced. Graph traversal is to be done separately.)

Combined changes:
 * 2013-08-05: Initial work on program & generator redesign.
 * 2014-10-05: Stagnant old program & generator redesign effort.
 * 2018-02-14: Clean-up, part 1. Add interpreter module.
 * 2018-02-16: Clean-up, part 2. Style change for new interpreter.
 * 2018-06-30: (Rebase.) Moved on top of "General clean-up"
               version.