2014-10-05: Redesign draft branch. (Non-functional.)
Non-functional draft for redesigning the
program module and audio generator. The
idea is to traverse the voice and operator
graphs before the generator, assigning
block IDs for the generator audio I/O
buffers before they are used.

The old design limits what the generator
is able to do, preventing full implementation
of the existing script syntax. Buffer IDs
properly pre-assigned would allow the
generator to run through a flat list of what
to generate, output buffers being filled and
re-used as inputs in the pre-scheduled order.

(The 2013 version of program.c re-added as an
extra file, the extra/earlier 2014 version in
the 2014 snapshot renamed.)

Change log:
 * 2013-08-05: Redesign sketch #1 (non-working)
               for program & generator.
 * 2014-10-05: Redesign sketch #2 (still non-working)
               for program.c, and minor change in parser.c.