2011-04-12: Rename project "sgensys". Redesign. Syntax changes.
Use per-block instead of per-sample processing for
sound generation.

Redesign parser-generated "program" for sound generator.
Use a sequence of "events", which are updates for the
sound generator to process, constructing and maintaining
its voices and operators accordingly.

The thinking was first to separate the old "program nodes"
into separate timed lists of "operator nodes" and "event nodes",
the former specifying what is to be used and the latter how.
But only "event nodes" were necessary.

Script syntax changes:
 - Added 's', silent time for operator (delay before sound).
 - Replaced center/left/right ('C', 'L', 'R') with 'b' for panning.
 - Added 'w' for changing wave type of operator.

Removed click reduction code in sound generator. It only
worked for basic wave forms (by fitting the duration of
a wave to the cycle length), not for modulated waves.