2011-04-10: "Event overhaul" redesign. Syntax changes.
Redesign parser-generated "program" for sound generator.
Use a sequence of "events", which are updates for the
sound generator to process, constructing and maintaining
its voices and operators accordingly.

The thinking was first to separate the old "program nodes"
into separate timed lists of "operator nodes" and "event nodes",
the former specifying what is to be used and the latter how.
But only "event nodes" were necessary as input.

Script syntax changes:
 - Added 's', silent time for operator (delay before sound).
 - Replaced center/left/right ('C', 'L', 'R') with 'b' for panning.

Removed click reduction code in sound generator rather
than fixing it for the new per-block processing. It only
worked for basic wave forms (by fitting the duration of
a wave to the cycle length), not for modulated waves.