2011-05-06: Add "VALueITeration" syntax, part 1. More.
Parameters can now be set to "[...]", with sub-parameters
deciding how the value changes over a specific time. This
provides minimal envelope support and is done ahead of
adding full-fledged ADSR support in the language.

Sub-parameters:
 - 's' (shape of change, only "l" (linear) currently done)
 - 't' (time to reach target value)
 - 'v' (target value)

Supported for amplitude, frequency, and panning.

Also change frequency ('f') to accept note values, both directly
and for 'v' inside '[]'. (This change broke note values for 'n',
which is now obsolete.)