2011-07-01: Voice & operator graph structure, part 1. More.
Redesign: Part 1 of changes to use an operator graph for each
voice, each operator specifying its modulators using an adjacency
list. Event data is also split into a main struct keeping track
of timing, linked to accompanying voice and/or operator updates.

See next snapshot for the completed (but not bug-free) version.

Script syntax: Rename '/' (delay time) to '\'. The old meaning
of '\' is scrapped, 's' (silence) providing what it did when
combined with '\'.