Split scriptconv stage from parseconv. Fix edge case.
Part 2 of redesign for parseconv & generator
(planned in 2013). Not exactly as originally
thought of, but as a basic preparation, split
parser output representation from script data
representation which is later converted to
the internal audio generation "program".

Fixes output for updates to operators created in
lists cleared in the same event. The updates are
now made to the right (unused) operator, instead
of to the wrong operator.

Design changes:
 * The parser is still mostly the same, with
   a very small simplification.
 * The old parseconv has become scriptconv,
   with minor adjustments for input data changes.
 * New parseconv mostly copies parser output,
   but leaves flat per-event operator list.

Todo: More refactoring, further redesign
(for which the door is now open).

Increase version to v0.2.15.

Change log:
 * 2019-10-16: Placeholder for separating stages.
 * 2019-10-23: Split parseconv module into
       parseconv and scriptconv.
 * 2019-10-25: Minimal cleanup.
 * 2019-10-26: Fill in more, still not functional.
 * 2019-10-27: Make functional.
       Set version to v0.3.3.
 * 2019-10-28: Fix 'test-builder' build.
 * 2019-10-31: Move down to after v0.2.14.
       Set version to v0.2.15.
 * 2019-12-12: In parseconv, remove unneeded typecast.
       In scriptconv, update commented-out operator re-use
       code, copied from voice re-use code. (Now compiles,
       but still not functional.)
 * 2019-12-22: Merge composite timing fix.
       Merge add missing default time adjustment
       for amp2 and freq2 ramps.