Fix nested scope checks for syntax elements.
Handle accepting/rejecting of parameters and
syntax elements according to nesting status,
in a way which depends on what's set for the
op instead of the current parse level (when
possible).

Corrects handling of 'r' for "@label" refs.

Todo: end_operator() is called too many times
for the same op; replace the temporary fix
which guards with a flag with something better.