Add help module, expand -h with <topic>.
Tweak usage notice for conciseness.
Change the -h option to list topics
available with -h <topic>. For now,
only other lists are available, the
first built-in help consisting of a
'wave' listing of recognized types.

Parser warnings now use the new code for
listing valid items. But a meta-list, of
the lists to be available with -h, needs
manual extension in help.c when more are
added. All lists are in the same format.

Change log:
 * 2020-01-18: Add help module, expand -h.
       Simple code for printing out name array,
       used for wave module. Easily extended to
       print out more lists in the same format.
 * 2020-02-16: Split out "Common options" usage line.