Add help module, expand -h with <topic>. Cli clean-up.
Tweak usage notice for conciseness.
Change the -h option to list topics
available with -h <topic>. For now,
only other lists are available, the
first built-in help consisting of a
'wave' type list and a 'ramp' list.

Parser warnings now use the new code for
listing valid items. But a meta-list, of
the lists to be available with -h, needs
manual extension in help.c when more are
added. All lists are in the same format.

Change cli argument parsing to allow a flag to be followed
by its argument without a space in-between. Recognize "--"
as meaning no further strings are flags. Uses SSG_getopt()
derived from Christopher Wellons's public domain getopt().

Change log:
 * 2020-01-18: Add help module, expand -h.
       Simple code for printing out name array,
       used for wave module. Easily extended to
       print out more lists in the same format.
 * 2020-02-16: Split out "Common options" usage line.
 * 2020-07-23: Move to saugns
       from the 2020-06 "mgensys" experiment.
       Replace 'noise' help type with 'ramp'.
   Old branch:
    - 2020-02-10: Add noise module with noise names,
          and add help type.
 * 2020-11-20: Move to ssndgen v0.2-dev.
 * 2020-12-20: List 'help' help type in itself.
       Change -h to also list normal usage info
       unless a topic is named.
 * 2021-01-08: Add SSG_getopt() and use in ssndgen.c.
       Change common module license from ISC to 0BSD.
       Add and use SSG_IS_ASCIIVISIBLE() from parser.
       Point the user to pass -h on cli option error.