Support AUDIODEV environment variable.
Check for the AUDIODEV environment variable,
using it to override the default device name
regardless of which type (ALSA, OSS, ...) is
used. Similar to handling in other software,
e.g. SoX and SDL. Empty strings are ignored.

For OSS only, also check OSS_AUDIODEV if the
AUDIODEV variable wasn't used, or "default".

Credit to Art Nikpal (@hyphop at GitHub) for
the basic idea, and SoX code as a reference.

Further refactor player/audiodev code a bit.