Simplify parser node lists and their use.
Remove the ptrarr module use in parser, and simply use
the linked list node structure for everything, with an
extension. Use mempool for the allocations. This means
no memory leaks even though e.g. adding and clearing a
list of operators in the same event leaves allocations
unused/dangling for those operators for the most part.

Change log:
 * 2021-07-17: New simpler work to use linked node lists from
       parser. Redone for "older" sgensys.
 * 2021-07-18: Debug and re-tidy, matching earlier clean-ups.
 * 2021-12-23: Merge changes to allocate and use
       temporary event branch structs. Add & use temporary
       mempool for allocations not kept in final output.