Simplify parser & parseconv node lists and their use.
Remove use of ptrlist module in parser, and simply use
the linked list node structure for everything, with an
extension. Use mempool for the allocations. This means
no memory leaks even though e.g. adding and clearing a
list of operators in the same event leaves allocations
unused/dangling for those operators for the most part.

Removes the SGS_POPP_ADJCS flag and parameter. Replace
the adjcs field and structure with one field for every
modulator type, using the same type for carriers. Tidy
up handling in parseconv a bit. Increases pointers, by
a few, but otherwise tidies up a bit in generator too.

Change log:
 * 2019-12-26: Remove SAU_POPP_ADJCS flag,
       replace checks using other information.
       Replace SAU_ProgramOpGraph and SAU_ProgramOpAdjcs
       types with SAU_ProgramOpList, simplify code.
 * 2020-10-07: Merge with changes from after new node list
       code was added and used, preserving old ptrlist usage.
 * 2021-07-17: New simpler work to use linked node lists from
       parser. Redone for "older" sgensys.
 * 2021-07-18: Debug and re-tidy, matching earlier clean-ups.
       Add 2020-12-04 devtests/pm-addremaddrem.sgs.
 * 2021-12-23: Merge changes to allocate and use
       temporary event branch structs. Add & use temporary
       mempool for allocations not kept in final output.