Refactor parser more & fix edge case. Revise mempool again.
Parser refactoring:
 * Make parser produce a linked list of modulator sublists
   per parent node. Simple and more generic than using one
   pointer per type of modulator list.

This fixes the test case of "devtests/ref-unused_node.sgs"
where referring back to an orphaned node in a script makes
the reference touch and modify the wrong node (because the
orphaned node hadn't had its ID properly set). Previously,
this was fixed by a larger change splitting the post-parse
processing stage, but the modulator list change is enough.

Changes to mempool module and its uses:
 * Style change in mempool.
   Shorter mempool method names.
 * Add mempool mpregdtor method.
   Use mpregdtor in symtab.

Change log:
 * 2021-07-23: Tidy parser code a bit more. Make modulator
       lists from parser a list of sublists; shorter code.
       Move down "devtests/ref-unused_node.sgs" test case.
   Old branch:
    - 2019-12-22: Add 'devtests/ref-unused_node.sau'.
 * 2021-07-24: Tidy parser yet more.
 * 2022-11-17: New mempool version,
       adding the mpregdtor method.