Rework '/' timing modifier.
Syntax changes:
 * A ";" after a "/" which splits a step will now
   catch the forward time shift, so that it's not
   added to events after the compound step. A "/"
   after a ";", but before a new step, applies to
   the new step like earlier. This makes ";" more
   like a barrier similar to "|" relative to "/".

This went along with solving currently known
bugs. But '/' is now quite different, design
must change to include time_durgroup() logic
to fix semantics for the node-split default.

Use same sub-event logic as for ';' and '\'.
Add timekeeping for dangling wait "carry" to
add to the next event, and subtract from the
duration.

As before, all new time-keeping takes place after
parsing, with only the setting of a flag added to
the main parser code.