Unify ';' and '\' syntax variations, use ';' in both roles.
Language changes:
 * Rename '\' with number to ';' with number. The same use.
   A side-effect is that ';' without number now needs to be
   used with no numerical expression characters just after.
   I.e. it now requires whitespace after, in various cases.

(A previous idea for a new syntax was to change ';' to ',',
but it looked less clear, prompting ';' to be added for end
of step, at least as an optional step terminator. This take
two simplifies, and keeps the spirit of, the older syntax.)

Adjust style in some scripts, since ';;number' is now nicer
to read. Also, rework 'examples/sounds/unnamed1.sau' a bit.

Adjust parsing, no longer end steps
at top level of script on newlines.

Change log:
 * 2022-05-22: Rename ';' separator
       to ','. Add optional ';' for
       explicitly terminating step.
 * 2022-05-23: Tidy code. Write new
       'Scopes' section on this and
       further in 'doc/README.SAU'.
 * 2022-05-24: Adjust scripts more.
 * 2022-06-08: Rename ',' separator
       back to ';' (remove new ';')
       and rename '\' to ';' (where
       the presence or absence of a
       number selects the meaning).
       Editing in 'doc/README.SAU'.