Improve script number reading precision.
Rework SAU_File_getd() to have less rounding error,
at very small size and speed cost. It is still less
rigorous (and far simpler) than system strtod() and
it still restricts how numbers can be written more.

Now, it is for example able to read digits of pi to
a fully accurate double. Before, the rounding error
was large enough that precision was reduced, almost
to single precision. But some numbers will still be
off a little as corner cases are still not handled.

Output after parsing is reduced to single precision
anyway, so the remaining flaws may not matter much.