Expand symtab code.
Add a second layer of data to the symbol table
beyond the string hash table. Planned features
will need it. Current uses are reworked to use
the new code in place of the old, with similar
result (only warnings from the parser differ).

The previous use of string pointer comparisons
is now replaced by lookup of symbol items from
a starting point provided by a string if there
are any items associated with it. There's only
name IDs and one type of variable yet, though.