Use first version of better 'm' mode 'g'.
I tried to approximate a soft-saturated
Gaussian noise, instead of first approximating
the real Gaussian noise, after realizing it's
easier.

Use two curves:
 * Polynomial approximation of mirrored
   and shifted Gaussian noise amplitude
   curve (the "sqrt(-2.0 * log(x))" part).
   Mirroring from the second half of the x
   domain makes it symmetric, well-defined
   at all values. But also too compressed.
 * A distortion curve which expands higher
   amplitude contrast while sucking down
   middle amplitude levels, is applied
   after to make the curve more realistic.

It sounds rather similar to Gaussian noise,
but a little softer. And the output amplitude
range is the full and bounded +/- 1.0.

The x values used are shifted so that input
somain is +/- 0.5, instead of 0.0 to 1.0.
This made it easier, and will also be used
later to apply a simple polynomial sin(x),
to make it even faster.