Make too-high frequencies behave the same on all builds.
Previously, on 32-bit x86, frequencies were saturated
to +/- half the sample rate, as a consequence of lrintf().
On x86-64 this didn't happen -- instead it wrapped around.

This is separate from the issue fixed in v0.3.12 affecting
PM -- a related but different lrintf() vs llrintf() issue.

Now, pick the wrap-around behavior as the standard; also a
cheap option to standardize on (using llrintf() if long is
less than 64-bit, lrintf() otherwise). In performance
terms, this is an insignificant change.

It's debatable whether saturation would actually be nicer,
though it would require more code and be slower (except on
e.g. 32-bit x86 if handled as special cases).

Frequencies resulting in values that don't fit in 64-bit
integers are still treated differently per platform, but
that's far less of a practical issue (scripts using such
extreme values may still not produce the expected noises).