Add 'uwh' line, for uniform white noise.
Use the ranfast32 PRNG to implement line
type mapping every position to a uniform
white noise sample: the "fill" and "map"
versions of the function differ a little
in details of the noise, but works well.

For sweeps, the result is a middle-value
plus noise, until the sweep is finished.

For 'R', the result is random chunks, of
selected length and frequency, where the
noise has a random amplitude and offset.