Add 3 line types with added white noise.
Language changes:
 * Add 'uwh' line (uniform white noise).
 * Add 'nhl' line (noise hump line).
 * Add 'ncl' line (noise camel line).

Use the ranfast32 PRNG to implement line
type mapping every position to a uniform
white noise sample; the "fill" and "map"
versions of the function differ a little
in details of the noise (it works well).

For sweeps, the result is a middle-value
plus noise, until the sweep is finished.

For 'R', the result is random chunks, of
selected length and frequency, where the
noise has a random amplitude and offset.

Also add 2 line types mixing a 'lin' and
some noise, in two different patterns. A
combination of mild noise and harmonics.
May or may not be useful to keep both 2.

License line module under LGPL.

Add 'examples/tests/line_noisy.sau'.

Change log:
 * 2023-02-21: Add 'uwh' line type.
 * 2023-03-10: Add 'nhl' and 'ncl'.