Rework and extend 'W' wave type collection.
Language changes:
 * Wave types.
   - Add 'ean' (evenangle -- Tillman's idea/name),
     add 'cat' (catear -- my own idea/name),
     add 'eto' (eventooth -- Tillman's idea/name),
     add 'par' (parabola -- the x^2 variety),
     add 'mto' (mellowtooth, my own idea/name).
   - Rename 'ahs' to 'spa' (sine parabola),
     adjust phase offset to match other wave types.
   - Rename 'hrs' to 'hsi' (half-rectified sine),
     adjust phase offset to match other wave types.
   - Remove 'ssr', was of less general use.

Organize wave types as 3 times 3 odd-even-all
added harmonics types after 'sin'; each group
has 3 types, in a mellow-medium-bright order.

With 'tri' go 'ean' and 'par', as the mellow.
With 'srs' go 'cat' and 'mto', as the medium.
With 'sqr' go 'eto' and 'saw', as the bright.

Additional types that don't fit in the scheme
come after, in addition; more could be added.

Change log:
 * 2023-02-28: Complete wave type changes and
       adjust scripts. To re-add 'ssr' later?
 * 2023-03-14: Make wave module license LGPL.