Add copy syntax ':label', to copy a labeled object.
Language changes:
 * Label syntax.
   - Add ":name" expression, for copying the object
     pointed to by "name". An operator copy will be
     inserted where the reference is placed, unlike
     with an "@name" reference which merely touches
     the original object. A copy has separate time.
   - Remove the deprecated syntax for num. variable
     assignment, "'name=".

Implemented with a very small addition in generator
along with extra flags and logic handled by parser.

A copy inherits (duplicates) state from the copied.

Todo: Needs a reworking of audio block use, so that
nested generators that end up shared on duplication
of the carrier run once, not once for each carrier.
The general solution is a deep clone of the object,
including linked objects too. The performant way is
instead to avoid that when possible, reusing audio.

Change log:
 * 2023-10-23: Add copy/clone from label syntax but
       without handling of any shared nested nodes.
 * 2024-09-15: Remove deprecated numerical variable
       assignment syntax "'name=".
 * 2024-10-26: Change new syntax from '*' to ':' to
       give room for '*' as an arithmetic operator.