Debug scanner char numbering, refactor in parser.
Scanner needed to create scan frames after a get,
not before, and yet after an unget restore before
a get, not after the get. Rework & fix the logic.

Now ungetting and regetting no longer loses track
of character numbers past e.g. a C-style comment.

Ungetting and regetting is still not supported in
combination with reading strings and numbers, for
correct position information, but it is not done.

Add sauScanner_getc_after() and use it in parser.