Extend modulator list syntax.
Language changes:
 * Writing two or more modulator lists in direct
   succession will concatenate their contents.
   "<X><Y>" has the same meaning as "<X Y>".
 * Modulator lists now append new generator
   objects upon assigning them to the given
   parameter rather than replacing the old; to
   get the old behavior, use the '-' prefix for
   the list assignment, writing "-<...>...".

This "backport" was trivial with the old ptrlist use in parser.
However, this ties appended lists to parser node state; ptrlist
in a node referred back to may be older than the previous state
with uses of labeled nodes, then resulting in appending to some
outdated list state for this reason. Later list designs fix it.