Add 'R' noise function 'a' for additive recurrence.
Language changes:
 * Signal generator types. `R` mode `m` noise functions:
   - Add `a` (additive recurrence, low-discrepancy a.k.a.
     quasirandom sequence) by default based on the golden
     ratio. Add mode subparameter `m.a` for changing the
     multiplier used to the fractional part of a number.
   - Rename `r` (uniform random, default) to `u`.

The new noise function 'a' applies a multiplier to
the half-cycle count in order to place the amplitude
of the half-cycle. The default value, based on the
golden ratio, gives a rich deep hum of sorts. Values
affect sound and, if wrap-around happens too seldom,
also the pitch. Some values give rise to DC, only.

Script changes:
 * Add 'examples/tests/addrec15rand.sau'.
 * Update 'examples/misc3-2pm_R.sau'.

Change log:
 * 2024-07-28: Rename 'R' noise function 'r' ("mr")
       to 'u' ("mu"). Add noise function 'a' ("ma").
 * 2024-07-30: Further work on noise function 'a'.
       Add 'R' parameter 'm.a' for changing the 'ma'
       multiplier, by default based on golden ratio.