Rework 32-bit PRNG seed handling in 'R' oscillator.
The old cycle & phase buffer filling code discarded
the highest cycle bit (used as PRNG seed and state)
for 2x mode. Merge code and use the same processing
loops for both 1x and 2x modes. The float constants
differ according to rate but no longer bitshifts...

The lowest bit must be discarded instead, the phase
displacing it when the rate is doubled. The code to
update parameters now toggles phase bits placement.
Seed differences using 1x and 2x modes are removed.

Fixes an inconsistency for overriding seed with the
's' parameter. Now for 'R' as 'N' it has full range
from 0.0 to 1.0 (or any value, modulo 1.0), instead
of the 0.0 to 0.5 result repeating, for 0.5 to 1.0.

This changes default 'R' instance seeds, scripts...