Add 'smo' line type (Smoothstep degree 5).
Language changes:
 * Line types.
    - Add `smo` (Smoothstep degree 5),
      a sinuous curve traditionally used for Perlin noise.

It sounds a bit brighter in a pleasant way when used for 'R'
instances producing noise more generally, so it makes sense
to offer as an option. (Waveshaping options would allow more
control but this curve is well-known and worth adding now.)

This curve also sounds pleasant enough to use for PM carriers
and modulators to brighten the sound a little, when wanted.

Rework example scripts:
 * Replace 'examples/misc3-2pm_R.sau'
   with new-old 'examples/sounds/babbly-metallic.sau'.
 * Add 'examples/sounds/bass-sounds2.sau'.
 * Tweak 'examples/sounds/bg-scape-01.sau'.
 * Tweak 'examples/sounds/cat-purr.sau'.
 * Rework 'examples/sounds/kaboom0.sau', rename 'kaboom1.sau'.
 * Tweak 'examples/sounds/music-elem-00.sau'.
 * Tweak 'examples/sounds/music-elem-01.sau'.