Expose/re-add naive 'W' using new mode 'm' toggle.
Allow use of `W` as a naive oscillator using a new mode `m`
switch. (The `W m` option is similar to `R m`, but simpler,
having only 2 letters for modes as yet.) The default is `a`
(ADAA, antiderivative anti-aliasing) preserving the current
behavior. Using `n` switches to the naive implementation in
the codebase that gives results like pre-v0.3.9 saugns, but
with the current wave types -- not rounded like v0.3.3 had.