Allow sweep and modulation for PD subparameter '.f'.
Not as useful as it optimally could be, because '.f'
still only shifts frequency properly when an integer
value is used. This limits uses of modulation giving
clean results. To be fixed by using accumulators for
rescaling of frequency later.

Refactor further for 'R' and 'W' implementations and
reduce duplication of code further.