Fix for builds with the clang 19 compiler.
This is for an issue known to have affected x86-64.

Unlike gcc and clang 18 and earlier, clang 19
implements lrintf() and llrintf() math functions
differently. [Update: This was deemed a bug in
clang/llvm, clang 20+ will not be affected.]

Always use llrint()/llrintf(), never lrint()/lrintf(),
to ensure that float values in 64-bit int range, when
converted to int, wrap around instead of saturating
when (later) assigned to 32-bit ints.

Outside of the clang 19 behavior, it used to be that
when `long` is 64-bit, lrint() and lrintf() sufficed,
but clang 19 may substitute 32-bit instructions for
these even then, whenever optimizing with -ffast-math,
unlike other compilers I'm familiar with.

The clang 19 behavior bugged out v0.5.0b and earlier,
preventing FM and PM and more from working correctly,
as phase offsets were saturated instead of wrapping
around in the expected integer arithmetic manner.