Add 'rs' stretch option for ADSR release stage.
Allow removing the "S" (sustain stage) in ADSR,
by stretching the release stage to fit the time
left. Writing `rs`, using the literal `s` added
for the release parameter only, switches it. To
get an ADR envelope instead of a AR envelope, a
sustain value can still be set -- the level the
decay reaches when the release stage continues.

Script changes:
 * Tweak  'examples/simple_sequence-pm.sau'.
 * Rework 'devtests/melody00.sau' into
          'examples/tests/tone_seq-metallic.sau'.
 * Tweak  'devtests/melody0-slow.sau'.
 * Rework 'devtests/melody0-fast.sau' and move to 'examples/'.
 * Rework 'devtests/melody0-slow.sau' and move to 'examples/'.
 * Move   'examples/simple_mixed_up-pm.sau' to 'devtests/'.

2025-10-09: Fix typo/bug in parsing of 'r' time literals.
The bug would prevent parsing of one-character constants.