Tweak envelope time handling, add more modes.
If times for an envelope together are too long for the
current "note" duration (i.e. the current `t` value for
the audio generator), handle this better than not ending
the trailing envelope stages that didn't fit the note.
The prior minimalistic code made the next note continue
in the midst of an old envelope cycle, not a new one.

It can also be interesting to use an envelope in a loop
configuration.

Three new behaviors are implemented (`ec`, `et`, `el` --
plus an `e0` mode to turn off the envelope), replacing
the old behavior. The default is the `ec` "clamp" mode.
 - Clamp envelope times to note duration. Reduce times
   so that the targets of the envelope stages are always
   reached (with or without any abrupt jumps if entire
   stages are reduced to 0 time). Times are shortened
   beginning at the last stage (release) and backward.
 - Truncate envelope shape if it doesn't fit the note.
   If e.g. the release is too long, the end value will
   not be reached, the trajectory left unfinished, but
   the stage is still finished so the next note starts
   (the level jumps) with the beginning of the attack.
 - Loop envelope. Sustain is given no time, and as soon
   as the release is done, the attack begins anew. This
   is an alternative to triggering an envelope per note.

The envelope is now also enabled with all times 0 for a
sustain level other than 1.0, unless `e0` is used. This
makes behavior consistent; a time of 0 otherwise means
instant change, and as a result of all times being 0,
the envelope turns into a gain control.

Refactor envelope parameters and their handling.

Script changes:
 * Tweak 'devtests/melody1-pm_vary.sau'.
 * Add   'examples/tests/env-wave.sau'.