Also reuse object IDs, not only generator IDs.
Cuts down ID count significantly, sometimes to
near or at that of the generator ID count. The
approach is less aggressive than for reusing a
generator ID, as moves made can't be undone...
 * Don't change ID for an existing object, but
   merely list the expired ID for reuse later.
 * Exclude all objects reachable from "labels"
   (names in scripts) including modulators, as
   these may be referenced at any later point,
   it can't be known in advance.

2025-10-16: Fix bug for cloned modulators. The
is_labeled state for some carrier "'a :a" with
modulators, though set for original modulators
was missing for the cloned modulators, causing
faulty object ID reuse.

2025-10-20: Avoid spurious graph traversals on
voalloc, for an update to a labeled generator.
This is a change to old behavior, which mainly
cleans up -p output, not printing some useless
updates. Check generator time param in voalloc
update and if generator is nested, requiring a
root generator's play time to set a new graph.

2025-11-09: Fix: "W[R t2 ; a0] t1 /1 W[N] t2".
Handling ID reuse needs to check has_next_ref,
in expiry traversal too; R is an example here.
Also mark/check generator root as no ID when a
node is orphaned (parent node is expired). The
voice allocation code avoids a bogus/bug voice
this way.