Debug label ref inside list from compound step.
Timing of the added event for the `@name` inside
such a modulator list was never correct. The event
used to be added before the outer compound step's
delay time. (Prior to v0.3.12, it was instead added
after with an additional delay time.)

Detect case of a label ref inside a modulator list
and handle it specially. (Also make sure that nodes
for added generators in the list after a `@name` in
the list are assigned the right (earlier) event, not
that for the `@name`. No misbehavior yet but may have
resulted later.)

Finally, also debug `{}` subscope behavior while at
it. Specifically, the combination with `;` wasn't
working correctly, now scripts thus far pass.

Script changes:
 * Add 'devtests/ref-nestcomp.sau'.
 * Tweak 'devtests/subscope3.sau'.