Refactor, use one big valr pointer array and enum set.
Instead of SAU_PSWEEP_* not containing PD parameters,
have one big SAU_PVALR_* which does. The SAU_PDD_*
enum with its IDs is still needed though. Adjust
data structures and their allocation, indexing.

This change is meant to make it easier to address the
set of value range parameters and line and envelope
subparameters, in a uniform way. This is in preparation
for further redesign.